@@ -6,18 +6,19 @@
     "homepage": "https://enqueue.forma-pro.com/",
     "license": "MIT",
     "require": {
-        "php": ">=5.6",
-        "enqueue/enqueue": "^0.8@dev",
-        "enqueue/null": "^0.8@dev",
+        "php": "^7.1.3",
+        "enqueue/enqueue": "0.9.x-dev",
+        "enqueue/null": "0.9.x-dev",
+        "queue-interop/queue-interop": "^0.7",
         "doctrine/orm": "~2.4"
     },
     "require-dev": {
         "phpunit/phpunit": "~5.5",
-        "enqueue/test": "^0.8@dev",
+        "enqueue/test": "0.9.x-dev",
         "doctrine/doctrine-bundle": "~1.2",
-        "symfony/browser-kit": "^2.8|^3|^4",
-        "symfony/expression-language": "^2.8|^3|^4",
-        "symfony/framework-bundle": "^2.8|^3|^4"
+        "symfony/browser-kit": "^3.4|^4",
+        "symfony/expression-language": "^3.4|^4",
+        "symfony/framework-bundle": "^3.4|^4"
     },
     "support": {
         "email": "opensource@forma-pro.com",
@@ -35,7 +36,7 @@
     "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
-            "dev-master": "0.8.x-dev"
+            "dev-master": "0.9.x-dev"
         }
     }
 }
